Update final_model.py
The accuracy after 25 epochs for this model is 0.9643 which is high enough to work with. The single output neutron paired with the sigmoid activation function results in the output being a number between 1 defined as:
output {x < 0.5 = 0,
             x > 0.5 = 1}
So, the output can be matched to the encoded values in the data, and new data can be classified in the same way.